Added basic trigger support

+++ Src/InterstellarPilotEditorV2/Assets/Pixelfactor/EditorV2/Scripts/Pixelfactor.IP.SavedGames.V2.Editor/EditorObjects/Scripting/Triggers/EditorTrigger_ScenarioTimeElapsed.cs
@@ -9,6 +9,6 @@
         /// <summary>
         /// The time before evaluation in seconds
         /// </summary>
-        public double Time = 2.0;
+        public float Time = 2.0f;
     }
 }
